requirements.txt
msedge-selenium-tools==3.141.3
selenium==3.141.0
urllib3==1.26.4
